moved business utility logic to entity
