+++ package.json
@@ -20,7 +20,7 @@
     "artillery": "^1.6.0-29",
     "artillery-plugin-fuzzer": "^1.0.1",
     "autoprefixer": "^9.7.3",
-    "aws-sdk": "^2.581.0",
+    "aws-sdk": "^2.584.0",
     "aws-sdk-mock": "^4.5.0",
     "aws-xray-sdk": "^2.3.6",
     "axios": "^0.19.0",
@@ -43,7 +43,7 @@
     "eslint-config-prettier": "^6.6.0",
     "eslint-plugin-cypress": "^2.7.0",
     "eslint-plugin-jest": "^23.1.1",
-    "eslint-plugin-jsdoc": "^18.4.1",
+    "eslint-plugin-jsdoc": "^18.4.3",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-react": "^7.17.0",
@@ -107,7 +107,7 @@
     "s3-zip": "^3.1.3",
     "s3rver": "github:ustaxcourt/s3rver#created_presigned_post_support",
     "sanitize-filename": "^1.6.3",
-    "serverless": "^1.56.1",
+    "serverless": "^1.59.0",
     "serverless-domain-manager": "github:codyseibert/serverless-domain-manager",
     "serverless-jetpack": "^0.7.0",
     "serverless-latest-layer-version": "^1.0.3",
@@ -115,7 +115,7 @@
     "serverless-offline-dynamodb-streams": "^3.0.1",
     "serverless-plugin-aws-alerts": "^1.3.1",
     "serverless-plugin-bind-deployment-id": "^1.1.1",
-    "serverless-plugin-git-variables": "^3.3.3",
+    "serverless-plugin-git-variables": "^3.4.0",
     "serverless-plugin-split-stacks": "^1.9.2",
     "serverless-plugin-tracing": "^2.0.0",
     "serverless-plugin-warmup": "^4.7.1-rc.1",
